refactor: extract markdown table renderer
